requirements.txt
Django==3.1.8
djangorestframework==3.13.1
gunicorn==20.1.0
django-filter==21.1